Upgrade to 14.1 fixes Misc correctness fixes for .NET Core 3.0

+++ Workers/DatabaseSyncWorker/AsyncConnectionExtensions.cs
@@ -5,7 +5,7 @@ namespace DatabaseSyncWorker
 {
     public static class AsyncConnectionExtensions
     {
-        public static void SendCommandFailure(this WorkerConnection connection, uint requestId, CommandErrors error)
+        public static void SendCommandFailure(this WorkerConnection connection, long requestId, CommandErrors error)
         {
             connection.SendCommandFailure(requestId, error.ToString("D"));
         }
